Moved the lib/scrapers directory into lib/server/scrapers and fixed the env var imports to use /static/private - Claude
