+++ app/src/lib/server/scrapers/_types.ts
