+++ app/src/lib/server/scrapers/fixtures/rule34-api.json
@@ -1,0 +1,26 @@
+[
+    {
+        "preview_url": "https:\/\/api-cdn.rule34.xxx\/thumbnails\/1769\/thumbnail_f8769b4d0ec93ee5c38207b1693afccc.jpg",
+        "sample_url": "https:\/\/api-cdn.rule34.xxx\/images\/1769\/f8769b4d0ec93ee5c38207b1693afccc.jpeg",
+        "file_url": "https:\/\/api-cdn.rule34.xxx\/images\/1769\/f8769b4d0ec93ee5c38207b1693afccc.jpeg",
+        "directory": 1769,
+        "hash": "f8769b4d0ec93ee5c38207b1693afccc",
+        "width": 700,
+        "height": 1000,
+        "id": 1915346,
+        "image": "f8769b4d0ec93ee5c38207b1693afccc.jpeg",
+        "change": 1709918501,
+        "owner": "metalizer",
+        "parent_id": 0,
+        "rating": "explicit",
+        "sample": false,
+        "sample_height": 0,
+        "sample_width": 0,
+        "score": 1042,
+        "tags": "2girls areola areolae arms_behind_back arms_tied_behind_back bangs bayonetta bayonetta_(character) beauty_mark before_sex big_breasts black_hair blonde_hair blush blushing bondage bound bound_arms breasts cherry-gig closed_eyes completely_nude completely_nude_female crossover dominant dominant_female drool drooling earrings erect_nipples eyeliner eyeshadow eyewear female female_only femdom femsub gag gagged glasses gloves hair_bondage hair_gag hair_grab hair_pull human human_only large_areolae large_breasts lesbian lezdom lezsub long_hair metroid multiple_females multiple_girls nintendo nipples nude nude_female painted_nails perky_breasts ponytail prehensile_hair restrained saliva saliva_trail samus_aran submissive submissive_female sweat thick_lips tied_hair tied_up very_long_hair witch yuri",
+        "source": "",
+        "status": "active",
+        "has_notes": false,
+        "comment_count": 13
+    }
+]
