+++ app/src/lib/server/scrapers/generic.ts
